change code product, variations function

+++ backend/database/migrations/2024_09_30_145307_create_combos_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('Cb_Sale');
             $table->integer('Cb_PriceAfterSale');
             $table->string('Cb_ImageUrl')->nullable(); // not checked
-            $table->unsignedBigInteger('CategoryID');  // foreign key
+            $table->unsignedInteger('CategoryID');  // foreign key
             $table->timestamps();
         });
     }
